+++ src/test/java/com/zettix/graphics/gjkj/vecstuffTest.java
@@ -9,38 +9,38 @@ import org.testng.annotations.Test;
 public class vecstuffTest {
     @Test
     public void testCross() throws Exception {
-        V3 one = new V3(1.0f, 0.0f, 0.0f);
-        V3 two = new V3(0.0f, 1.0f, 0.0f);
+        V3 one = new V3(1.0, 0.0, 0.0);
+        V3 two = new V3(0.0, 1.0, 0.0);
         V3 cross = vecstuff.cross(one, two);
-        V3 expected = new V3(0.0f, 0.0f, 1.0f);
-        Float d = vecstuff.distanceSquared(cross, expected);
+        V3 expected = new V3(0.0, 0.0, 1.0);
+        Double d = vecstuff.distanceSquared(cross, expected);
         Assert.assertTrue(d < 0.0001, "Not small: " + d);
     }
 
     @Test
     public void testDot() throws Exception {
-        V3 one = new V3(1.0f, 0.0f, 0.0f);
-        V3 two = new V3(0.0f, 1.0f, 0.0f);
-        Float d = vecstuff.dot(one, two);
-        Float expected = 0.0f;
+        V3 one = new V3(1.0, 0.0, 0.0);
+        V3 two = new V3(0.0, 1.0, 0.0);
+        Double d = vecstuff.dot(one, two);
+        Double expected = 0.0;
         Assert.assertEquals(d, expected);
     }
 
     @Test
     public void testDistanceSquared() throws Exception {
-        V3 one = new V3(1.0f, 0.0f, 0.0f);
-        V3 two = new V3(0.0f, 1.0f, 0.0f);
-        Float d = vecstuff.distanceSquared(one, two);
-        Float expected = 2.0f;
+        V3 one = new V3(1.0, 0.0, 0.0);
+        V3 two = new V3(0.0, 1.0, 0.0);
+        Double d = vecstuff.distanceSquared(one, two);
+        Double expected = 2.0;
         Assert.assertEquals(d, expected);
     }
 
     @Test
     void testAdd() {
-        V3 one = new V3(1.0f, 0.0f, 3.0f);
-        V3 two = new V3(0.0f, 1.0f, 2.0f);
+        V3 one = new V3(1.0, 0.0, 3.0);
+        V3 two = new V3(0.0, 1.0, 2.0);
         V3 out = vecstuff.add(one, two);
-        Assert.assertTrue(vecstuff.distanceSquared(out, new V3(1.0f, 1.0f, 5.0f)) < 0.0001, "too far");
+        Assert.assertTrue(vecstuff.distanceSquared(out, new V3(1.0, 1.0, 5.0)) < 0.0001, "too far");
     }
 
 }
